Address review comments
 - Split properties into DockerComposeFactoryProperties - static, used by the DockerMachine/Compose initialization and DockerComposeTestPropertes - Spring properties used by the tests
 - Improve analytic test waiting logic. Instead of fixed time wait use the Awaitability library to wait until the metrics is send to the TSDB
